WOWL GPIO pin change 6->8
:Release Notes:
WOWL GPIO pin change 6->8

:Detailed Notes:
After WOWL On and  DC off and module was plugged out and in,
TV is turned on. GPIO6's default value is High, so TV is turn on like
WOWL. To solve this, GPIO is changed from 6 to 8. And for supporting
WEE board and new board with new module, WOWL is operated by both GPIO
in the same time.

:Testing Performed:
check on K6LP/WEE

:Issues Addressed:
[RTKWBTV-1288][LGE][RTL8822CU] When plug the module, TV has turn on by DC on.
[RTKWBTV-1323][RTL8822CU] WOWL GPIO change

Change-Id: Id4bd7c567d48ae90dca8aa01c0e564c809947683